Util: Move interface functions to controllerInterface.go
